api/requirements.base.txt
daphne~=4.1.2
django-filter~=24.2
djangorestframework~=3.15.2
drf-spectacular~=0.27.2
markdown~=3.6
django-guardian~=2.4.0
djangorestframework-guardian2~=0.6.0
